[test] Unit test change: handler->read/write
