+++ src/main/java/ssd/SSDConstant.java
@@ -1,6 +1,10 @@
 package ssd;
 
-public class SSDConstant {
+public static class SSDConstant {
+    public static char READ = 'R';
+    public static char WRITE = 'W';
     public static String SSD_FILE_PATH = "ssd_nand.txt";
     public static String OUTPUT_FILE_PATH = "ssd_output.txt";
+    public static int MIN_LBA = 0;
+    public static int MAX_LBA = 99;
 }
